+++ demo/src/scenes/demo/index.tsx
@@ -1,1 +1,0 @@
-export { default as DemoScene } from './demo'